Update frontend dependencies
This commit provides several frontend updates including adding new dependencies such as autoprefixer, postcss, sass, and tailwindcss to the package.json file. This allows improved styling and compatibility across different browsers for the frontend application.

The App component's JSX has also been greatly modified, adding new elements and improving the structure of the existing ones. Two new files - buttons.js and buttons.css - have been added which include styling for buttons. In addition to this, the style.css, buttons.scss and package-lock.json files have been updated with the inclusion of new styles.

In the backend file app.go, code was added to handle IPv6 addresses correctly by surrounding them with brackets. This is based on the condition that if the string contains a colon, it is classified as an IPv6 address. This ensures that IPv6 addresses are displayed and processed correctly.

These changes enhance the visual aesthetics of the frontend and ensure correct handling of IPv6 addresses in the backend.