New: `daisyui`; Remove unused buttons.js and create new UI effects
This commit removes an unused file, buttons.js and introduces a new glow effect on both CSS and Javascript files. The replacement and deletion of buttons.js are due to refactoring the UI, replacing the old button style with a glow effect for a more appealing user experience.

Frontend dependencies are also updated with the addition of 'daisyui', a plugin for tailwind CSS which provides a set of UI components. This aids in implementing the glow effect.

In App component, the implementation of loading spinner while terminating a process is added for better UI.

In the backend, unnecessary code to handle IPv6 addresses is removed in app.go. This was previously wrongly included and the task is dealt in a different portion of the application.

Lastly, package.json, package-lock.json, and tailwind.config.js files update due to the added 'daisyui' component.