Refactor frontend UI and improve style
The commit is a significant refactoring of the user interface in App.jsx to improve readability and functionality.

The 'overflow: hidden' CSS property of the glow.css was removed to explore a better way to deal with overflows. Also, hefty blocks of repeated JSX code in App.jsx were replaced with a reusable 'Item' component for clean-up and efficiency.

Furthermore, lots of new colors were added into 'extend' block of tailwind.config.js for giving more flexibility in terms of design.

A new Process termination feature was added in App.jsx to allow users to terminate running processes.

Lastly, a screenshot was added to the README.md for providing a quick preview of the app.